Finished the rest of the project.  Particle Rod Shot does not work properly and will confer with teammate on what the problem/solution might be

+++ SnazzyProject/Assets/Scripts/NonStatic/Particle2DRod.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class Particle2DRod : Particle2DLink
 {
 	public int mLength;
-	public int mRestitution;
-	public int mPenetration;
+	public float mRestitution;
+	public float mPenetration;
 
 
 	public override void CreateContacts(List<Particle2DContact> mContacts)
